requirements.txt
amqp==5.0.5
appdirs==1.4.3
astroid==2.5.1
autopep8==1.4.4
billiard==3.6.3.0
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==2.2.2
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-javascript-settings==1.1.2
django-login-required-middleware==0.5.0
django-model-utils==4.1.1
django-picklefield==3.0.1
django-timezone-field==4.1.1
gunicorn==20.0.4
html5lib==1.1
humanize==3.2.0
idna==2.8
importlib-metadata==3.6.0
ipaddr==2.2.0
isort==5.7.0
kombu==5.0.2
lazy-object-proxy==1.5.2
lockfile==0.12.2
mccabe==0.6.1
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.1.0
progress==1.5
prometheus-client==0.8.0
prompt-toolkit==3.0.16
pycodestyle==2.6.0
pylint==2.7.2
pyparsing==2.4.6
PyPDF2==1.26.0
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.4
pytoml==0.1.21
pytz==2021.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
typing-extensions==3.7.4.3
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.0
